adding the configuration and elimante Dao
