+++ security_module/src/main/java/com/apipharma/security/config/ApplicationConfig.java
@@ -1,4 +1,59 @@
 package com.apipharma.security.config;
 
+
+import com.apipharma.security.repository.UserRepository;
+import lombok.AllArgsConstructor;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+import org.springframework.security.authentication.AuthenticationManager;
+import org.springframework.security.authentication.AuthenticationProvider;
+import org.springframework.security.authentication.dao.DaoAuthenticationProvider;
+import org.springframework.security.config.annotation.authentication.configuration.AuthenticationConfiguration;
+import org.springframework.security.core.userdetails.UserDetailsService;
+import org.springframework.security.core.userdetails.UsernameNotFoundException;
+import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
+import org.springframework.security.crypto.password.PasswordEncoder;
+
+@Configuration
+@AllArgsConstructor
 public class ApplicationConfig {
+
+    private final UserRepository userRepository;
+
+
+    @Bean
+    public UserDetailsService userDetailsService() {
+        return username -> userRepository.findByEmail(username)
+                .orElseThrow(()-> new UsernameNotFoundException("User not Found"));
+
+    }
+
+    // data acess object that is responsible
+    @Bean
+    public AuthenticationProvider authenticationProvider() {  //  Data Access Object
+        final DaoAuthenticationProvider authProvider = new DaoAuthenticationProvider();
+        authProvider.setUserDetailsService(userDetailsService());
+        authProvider.setPasswordEncoder(passwordEncoder());
+        return authProvider;
+    }
+
+
+
+    @Bean
+    public PasswordEncoder passwordEncoder() {
+        return new BCryptPasswordEncoder();
+    }
+
+
+    @Bean
+    public AuthenticationManager authenticationManager(AuthenticationConfiguration config) throws Exception {
+        return config.getAuthenticationManager();
+        /* I will call it authentication manager and within  this mean I want to inject an object  of type authentication configuration, and I will call it config this authentication  configuration hold already the information about 1:40:04 the authentication manager so I will just return  config.getAuthenticationManager() all right
+        */
+
+    }
+
+
+
+
 }
