--- security_module/src/main/java/com/apipharma/security/dao/UserDao.java
@@ -1,35 +1,0 @@
-package com.apipharma.security.dao;
-
-import org.springframework.security.core.authority.SimpleGrantedAuthority;
-import org.springframework.security.core.userdetails.User;
-import org.springframework.security.core.userdetails.UserDetails;
-import org.springframework.security.core.userdetails.UsernameNotFoundException;
-import org.springframework.stereotype.Repository;
-
-import java.util.Arrays;
-import java.util.Collections;
-import java.util.List;
-
-@Repository
-public class UserDao {
-    private final static List<UserDetails> APPLICATION_USERS = Arrays.asList(
-            new User("azolis92@hotmail.es"
-                    , "password"
-                    , Collections.singleton(new SimpleGrantedAuthority("ROLE_ADMIN")))
-
-            , new User("user.mail@gmail.com"
-                    , "password"
-                    , Collections.singleton(new SimpleGrantedAuthority("ROLE_USER"))
-            )
-    );
-
-    public UserDetails findUserByEmail(String email){
-        return APPLICATION_USERS
-                .stream()
-                .filter(u->u.getUsername().equals(email))
-                .findFirst()  // Not repeated email;
-                .orElseThrow(()-> new UsernameNotFoundException("No user was FOUND")
-                );
-
-    }
-}
